Acq: fix widget choice for user-linked fields in Acquisitions Search
The Acquisitions Search interface (which also sits underneath the
Purchase Orders inteface, My Selection Lists, and one or two others)
stopped presenting the correct widget for search terms linked to
actor.usr.  This was brought on when PCRUD control was added to
actor.usr, so that means the bug affects only master, not rel_2_2 or
earlier.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>