Avoid CAPTURE-blocked holds in pull list (IDL view)
Update the new IDL holds pull list view to avoid returning holds for
users that have CAPTURE penalties.

Note that items are blocked from the pull list only if the hold pickup
lib is within the CAPTURE penalty org unit range.  This means that users
could be blocked at one branch, but use a pickup lib at another to avoid
the block.  To prevent this type of abuse, set the org_depth to 0 on any
penalties that apply the CAPTURE block.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>