kpac : repair Top link
Template Toolkit string.replace() does an inline replacement instead of
returning a value.  Use .replace() correctly when building the Top URL,
or the result will be an empty href, which direct to "here" in
Firefox/Chrome, but directs to "here, minus the page name in the URL"
(e.g. http://domain/eg/kpac/) in IE, which ultimately directs the user
to /eg/opac/login (since it doesn't match any configured paths).

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>