Move Prefix field ahead of Names in patron editor
In earlier versions of Evergreen (1.6 and earlier), the Prefix field
was used to enter title information, such as Mr., Mrs., etc. When the
field was eventually added to the newer patron registration screens
(2.0+), it was added before the patron's last name field. Ideally,
the prefix should be listed prior to the patron's first name instead.

In addition, address label issues by changing "Suffix/Title" back to
only "Suffix" and making "Prefix" into "Prefix/Title".

Note: string changes would normally not be allowed in a stable release,
however in this case the previous translation (if any) will still get
pulled from the IDL for the field in question, so English will not
suddenly start showing up in the patron registration form.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: James Fournie <jfournie@sitka.bclibraries.ca>
Signed-off-by: Dan Scott <dscott@laurentian.ca>