TPAC: Search wrapper spacing
Currently the search box options are "smooshed up" against the
links toolbar.  The following changes address this.

Padding was added to style.css search_box_wrapper in order to
bring the "Search the Catalog" down lower from the link toolbar.

Padding was also added to the searchbar.tt2 in order to bring the
search filter options down lower from the "Search the Catalog"
wording.

Signed-off-by: Melissa Lefebvre <mlefebvre@biblio.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>