Ability to add records to permanent bookbags in TPAC.
Add an actor usr setting for holding the name of a default book
list: opac.default_list.

Add a button on the my list interface to choose a list should as
the default list for adding titles.  There is presently no way to
unset a default menu, though the default can be changed to any
other menu at any time by clicking that list's button.

Modify opac/record and opac/results so that if a patron is logged
in their my lists will populate a menu.  This menu will include
options to add to a temporary list, the default list (if any), up
to 10 of the user's other lists, to add to a newly created list,
or to see all of their lists and add to one that does not appear
on the menu.

Adding to a temporary list will function the same as adding to a
list does prior to this enhancement.

Adding to a list chosen from the menu will add the record to that
list, and return the user to the search results or record page
that they were looking at.

Choosing to add to a new list will take the user to their "my
lists" page where they can create a new list.  After the list is
created, the record they wanted to add will be added to the new
list and they will be returned to the search results or record
that they were looking at.

Choosing the "see all" menu option will also take the user to
their "my lists" page.  However, all of their bookbags will be
visible and not just the normal limit of 10.  Their will be a
button next to each list's name with the text "Add to this list."
When the user clicks one of those buttons, the record will be
added to that list and the user's session redirected back to
their search or result page.

The user will have the option to create a new list when viewing
all of their lists.

The user will also be able to use the "Add to this list" feature
when they have chosen the "Add to new list" menu option.  This is
done from simplicity in the design, but also allows the user to
change their mind at the last second.

If a patron is not logged in, the add to my list will appear the
same as it does prior to this development.  It will continue to
function as it does prior to this development.

Add coust for opac.patron_temporary_list_warn to let an ou enable
warning users when adding records to a temporary book bag.

Add cust opac.temporary_list_no_warn to let patrons disable
the warning for themselves.

Add the opac.default_list usr_setting_type. This will be used to track a
user's default bookbag.

Add the upgrade script to create new coust and custs.

Insert coust (opac.patron.temporary_list_warn) and cust
(opac.temporary_list_no_warn, opac.default_list) into the respective
tables.

Add release notes for add to permanent bookbag feature.

Add a warning when the user adds a record to a temporary list.

TPAC has been modified so that a user will see a warning before adding
a record to a temporary bookbag.  This message serves to inform the
user that they are adding to a temporary list that will disappear when
their session ends.

A new org. unit setting has been added,
opac.patron.temporary_list_warn, that will enable this warning when
set.  Sites may choose not to display this warning.

The user may also set a preference in their search preferences to
disable this warning.  The setting only works when a user is logged
in, of course.

Add release notes for the temporary list warning feature.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>