TPAC: Set autofocus appropriately for different contexts
At login, the basic search bar was fighting (and winning) for the
default focus rather than the username field. Additionally, in the
advanced search pages, no field was getting focus. Therefore, check to
see if we expect to be in a context where it makes sense to set the
default focus to the basic search input field; otherwise, set the focus
to the login username field or the appropriate advanced search field.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>