ACQ : lineitem recovers focus from various interfaces
1. When viewing an invoice then clicking on the title link for a lineitem,
the user is taken away to the copy grid within the PO page for the
lineitem.This adds a "Return to Invoice" button in the copy grid which
takes the user back to the invoice, focused on the lineitem in question.

2. Adds a "Return to Search" button for PL/PO pages to return to the
lineitem in question when the PL or PO links are selected for a lineitem
in a set of search results.

3. Adds a "Return to Invoice" button to PO/Picklist pages with lineitem
focus when the page was accessed from the invoice UI.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>