TPAC: Slightly more accessible user-visible account notes display
Rather than a table layout that resembles an unordered list, use a real
table with table header so that screen readers will know what to
associate each column with.

Also, move the table display into its own class so that we don't pick up
the uppercase transform, and move it out of the div box so that we can
use whatever screen real estate is given to us, rather than forcing
notes into the hard-coded 662px box.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Bill Erickson <berick@esilibrary.com>