Fix / improve bookbag list paging
The prior code used CSS to hide elements, which is a bit weird given
that we can simply avoid creating the element in the first place and
keep the DOM a lot simpler.

Also, the previous code was subject to some bugs if it wasn't given
an explicit offset GET param or paging at the end of a list
(particularly given that the LIMIT parameter ensured that the number of
bookbags returned would always be equal to or less than the LIMIT
parameter).

Finally, the changes related the permanent bookbags didn't pass through
one of the context parameters we needed (bookbag_count) to enable
paging.

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Ben Shum <bshum@biblio.org>