ACQ repair reverse exchange rate in fund transfer
Updated fund transfer code to use the DB exchange rate stored procedure.
In addition to code consolidation, this allows the fund trasnfer
operation to work if an exchange rate for the two currencies involved is
only defined in one direction.  The existing code required that the rate
be defined in both directions in the database or it would produce an
error in the logs and no visible error in the interface.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>