TPAC: Fix org_unit location detection for login
The TPAC login function currently looks for a 'loc' param to pass
to the login service.  This isn't right because:

A. 'loc' is a search-related param, and shouldn't affect logins
B. The login form doesn't actually pass through the 'loc' param
C. The 'loc' param has been more or less replaced by 'locg'

No, this isn't multiple choice, it's all of the above.  Our best
option at this point is to use the 'physical_loc' value instead.
This matches what was done in JSPAC (though it was called 'ol' for
"original location" back then).

Finally, this commit is 95% the work of Dan Scott, including the
additions of logging info and a few whitespace fixes.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Dan Scott <dan@coffeecode.net>