Fix LP949249, Edit, then Create button creates a new item in all cases
This fix also resolves the following:
* https://bugs.launchpad.net/evergreen/+bug/806585,
Item Attributes saved on Close
* https://bugs.launchpad.net/evergreen/+bug/644753,
check for duplicate barcode during a create session in copy/volume editor

1. The Copy Editor screen is launched from the Volume Copy Creator,
which is launched from the Holdings Maintenance screen. If the default
interface is set, CE is shown in a separate window; if the 'unified'
interface is set, CE is embedded in the same window as VCC. Upon the
user clicking Save or Close in the Copy Editor, the intent is to create
a new item record or not, respectively.  However, it happens that a new
record is always created, because VCC instructs CE to let VCC update the
item, which will occur whether the user has clicked either button.

To fix 1, if the default interface is set, we reconfigure CE to update
the item or not, depending on the button clicked; otherwise, in the
unified interface, we allow VCC to update the item or not.

2. Upon updating the item, if the barcode is a duplicate, the ensuing
sequence of events is problematic: a database-level alert message is
shown, Copy Editor is closed, and Volume Copy Creator is closed, thus
forcing the user to re-formulate the request from Holdings Maintenance.

To fix 2, we show a higher-level alert message to the user. If the
default interface is set, we stay within CE, reconfigure the barcode to
be an input field, and focus the user on it; otherwise, in the unified
interface, we return the user to an area of the screen where the barcode
field is available. In either case, the user can re-enter a barcode
value and re-submit without losing any values previously entered into CE.

3. We also take the opportunity to enhance the coding performance of
Copy Editor, by delegating the handling of the button and key clicks to
the parent groupbox element. Previously, the event handler was attached
to several child elements, which is not performant.

Conflicts fix:

	Open-ILS/xul/staff_client/server/cat/copy_editor.js

Signed-off-by: Ben Shum <bshum@biblio.org>