Allow propagation of encumbrances without funds
Some libraries must return all unspent money to some funding agencies at the
end of the year.  To support this, we give the year-end close-out operation
the ability to dump the money from rollover-enabled funds into the void. Funds
are still created for the next year, and encumbrances are moved to theses
analogous funds.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>