Get the TPAC basic link inline for <noscript>
The TPAC basic link appeared outside the <p> tag inside the <noscript>
section of the JSPAC, meaning that it would display on the far left
border of the screen rather than right after the colon. Fix that by
getting it inline.

Also apply a style to the link so it appears as some kind of a link.
What kind of madman overrode the basic <a> element style to make it
appear as plain text by default in the first place? :P

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>