Trying to update patron CSS color fails here, so don't try
After modifying user standing penalties ("messages" in the staff client
patron interface), you could get this error message in an alert() dialog:
    TypeError: patron.display is undefined.

We were apparently trying to use patron.util.set_patron_css() wrongly,
so the commented out changed verison of the call got farther, but stopped at
    ...penalties[i].standing_penalty().name is not a function.

Which probably has something to do with unfleshed data.  Somebody can
get to the bottom of this if they want, but for now it's easier just not
to try to do this and make the alert() dialog go away.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>