Templatize TPAC styles
Begin by gathering the colors together in one consistent file, such that
skinning for the purposes of color only needs to touch one small file.
Also, it should make things easier for designers to reduce drastically
the number of colors that are actually used in the design.

Also, enable TT to handle the CSS files in Apache configuration.

Also, convert the login background image into a simple gradient.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>