Add pub date to CSV output.
Change the template for the ContainerCSV reactor to add the publication
date (MARC 260$c) in the csv output in the same position that the previous
commit adds it to the html output.

Add an upgrade script to update the existing template for the ContainerCSV
reactor in action_trigger.event_definition.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>