Vandelay Copy Overlay support
Vandelay Item Attributes (Cataloging -> MARC Import /
Export -> Import Item Attribute Definitions) contains
a new field called "Overlay Match ID".  The presence of data
in this field extracted from an import-item copy indicates
to the Vandelay import process that a copy overlay is requested
instead of new copy creation.  The value for the field is the
copy id for bib record queues and the ACQ lineitem_detail ID for
Acquisitions Queues.  For either type of queue, however, overlay
occurs against a real copy (asset.copy).  In the ACQ queue case,
we use the lineitem_detail ID because this is the data ACQ
providers and sub-systems will have access to.

When a match point ID value is a set and a matching copy is found,
the values extracted from the inbound copy data are used to replace
values on the existing found copy, including the call-number label.
Any fields on the inbound copy that are empty are ignored.

One use case for this feature are shelf-ready items produced by a
3rd-party (e.g. ACQ provider) and delivered to the library via MARC
file for upload.  The file might contain improved MARC bibliographic
data as well as real barcodes (i.e. not temporary ACQ generated
barcodes) for the copies already purchased through the vendor.

Permission
~~~~~~~~~

This adds a new permission called IMPORT_OVERLAY_COPY which is
required to perform the copy overlay step.

Regardless of permission, it is not possible to overlay values on
a copy unless the imported bib record links (creates/overlays/merges)
to/with the owning bib record for the copy to be overlaid.  This is
both for security and removal of a potent foot-gun.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>