ACQ unified search submit on enter (for real)
This replaces the old-style submit on enter handler which was based
attaching key handlers to the dijit domNode, which does not work for
filtering selects (i.e. more complicated dijits) w/ the dijit-sanctioned
dojo.connect(..., 'onkeyup', ...), which works regardless.  This also
covers more form input elements which were overlooked with the original
submit-on-enter code.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>