Add dependencies for Fedora
Noted by Net Wolf on the mailing list, some core dependencies were
missing from the Fedora Makefile.install installer. As some of them have
recently been packaged for Fedora 17, add them to the package
collection; others are still unpackaged, so leave them there.

Note: the packages are not available on Fedora 16, but that has a
limited remaining shelf life anyway, so consider this a shift to
supporting only the latest stable version of Fedora.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>