TPAC: Add the ability to set a different default locale
Installations with multiple sites might desire different default locales
for their skins. This adds the Apache config directive
"OILSWebDefaultLocale", which enables you to set the default locale on a
global and a per-skin basis.

Note that this prevents the browser Accept-language header from having
any effect; sadly that header is not often used in real life.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Art Rhyno <art632000@yahoo.ca>

Conflicts:
	Open-ILS/examples/apache/eg_vhost.conf

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>