TPAC - fix pickup locations for edit holds
When a patron goes to edit a hold in TPAC, they have the option to
choose a different pickup location. Unfortunately, it looks like that
dropdown does not prevent them from selecting locations that aren't
normally allowed for pickups, such as system groups or consortium
itself.

This patch adds a check for 'can_have_vols_only' to the edit holds
screen.

Credit to Melissa Lefebvre for finally nailing down the source of
the issue.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>