Allow some buttons in Import via Z39.50 to wrap lines
Some of our technicians with 4:3 ratio monitors and a reasonable
font-size are unable to see the "Import Record" button on the Import via
Z39.50 screen, because the buttons in that row are laid out in a single,
non-wrapping row, and there is no affordance for scrolling over to the
side.

Simply changing the containing "hbox" element to a "description" element
allows enough of the buttons to overflow to resolve the problem in our
experience.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Bill Erickson <berick@esilibrary.com>