Simplified Hold Pull List: Fix several sorting bugs
First of all, sorting on most columns was broken due to a bug in the way
that the flattener methods of the open-ils.fielder service were
constructing their SQL JOINs.  We were coming up with way too many
joins, and then losing track of which JOIN's alias to refer to when
building the ORDER BY clause later.  This is fixed.

Secondly, the shelving location column now sorts automatically by the
shelving location *ordering* values, when avaiable.  These are the
values that you set up in the drag-and-drop staff client interface
titled "Copy Location Order."  When these values are not set for the org
unit whose pull list you're viewing, the sorting will fall back to
alphabetical.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>