TPAC: Reset Form and New Search should clear queries
Currently, clicking "Another Search" in the search results page, or
"Reset Form" on the advanced search page, keeps the same query terms and
type around. Clear these parameters as expected; essentially, we keep
only the search location (loc or locg) in scope.

Note that we explicitly keep parameters around when someone is going
from a basic search to an advanced search, because the assumption is
that they want to refine their search further.

Also affected are the org_selector, because we want to pull the logic
for loc or locg out into a common area (the get_library block in
misc_util.tt2), and the style (input.opac-button elements now get styled
the same as a.opac-button elements, which makes it easier to intermingle
them on the same form).

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>