TPAC - fix alignment of "You are now browsing..." in shelf browser
The "You are now browsing..." message does not align to the center of the
page. Changing the <div> to a <p> seems to properly align the text.

Along the way, we should remove the extra <span> and <strong> (the strong
is redundant with the css being inherited by color_4 class) and make the
new single line message more i18n-friendly for future translations.

Credit to Michael Peters for noticing the problem and thanks to Dan Scott,
Lebbeous Fogle-Weekley, and Thomas Berezansky for help via IRC for the
i18n-bits.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Michael Peters <mrpeters@library.in.gov>
Signed-off-by: Dan Scott <dscott@laurentian.ca>