Set a sane initial MFHD Editor size
For most MFHD records, the size of the MARC Editor window when spawned
fits into the confines of a normal desktop. However, for MFHD records
with dozens or more rows, the window chrome can be inaccessible because
it is bumped off the top of the Windows desktop. This makes it
impossible for users to close the window without a workaround like
hitting ALT+F4

This branch sets the window to an initial height of 40 pixels less than
the height of the entire staff client window. This should be enough to
enable the chrome on top to be visible. In addition, we have to set the
width to something reasonable because XUL seems to otherwise set the
window to be extremely skinny once an explicit height is set, so we set
that to be half the width of the staff client window.

Finally, we allow the new window to be resizable according to the user's
preference.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Mike Rylander <mrylander@gmail.com>