Authorities: Fix broken UI due to new XUL/Dojo/regressions
New versions of XUL, since Firefox 10, hate dojo.query("foo bar"), so we
have to refactor it to use dojo.query("foo").query("bar") instead. At
least until we upgrade to a new version of Dojo.

In addition, 522d8d82 introduced a regression meaning that the text of
the authorities was no longer wrapped in a SPAN element with a class of
"text". That in turn meant that the Dojo queries weren't picking up the
labels for the merge box, leaving no visual indication to the user as to
what authorities exactly were being merged. So, repair that.

Also, similar to a fix for the MFHD editor, set the initial height of
the spawned MARC Editor for an authority record to something reasonable.

Signed-off-by: Dan Scott <dscott@laurentian.ca>

Conflicts:
	Open-ILS/web/js/ui/default/cat/authority/list.js

Signed-off-by: Mike Rylander <mrylander@gmail.com>