Simplified Hold Pull List: Fix broken filters
Only two filterable fields were available, and at least
one didn't even work.

Now we have a few more, and they seem to all work for me in testing.

This problem was reported to me by Thomas Berezansky.

While we're at it, also make sure the list of shelving locations in that
filter dropdown is sorted alphabetically.  This was Kathy Lussier's
request.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>