Fix sorting when printing from FlattenerGrid-based interfaces
This particularly affects the Simplified Hold Pull List.  The printed
verison of your data is supposed to reflect the same sort order that has
been selected in the visual interface, and now it does.

Buggy handling of the different avenues of sorting the grid (clicking column
headers, using the column-picker/multicolumn-sort dialog, default sort order)
had broken the flow of the data needed to build the right sort clause for
some situations.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>