LP#1410532: TPAC - Make "Show more details" in results optional
Academic sites typically handle a lot of electronic content and want to
bring most details about entries, such as availability, status, and
URLs, on a results screen to the front.

This adds a 'show_more_details.default' option in 'config.tt2' that can
have the values 'true', 'false' and 'hide'.  The 'true' and 'false'
values set the default for the button to more and less respectively, and
'hide' suppresses the button from showing at all.

Signed-off-by: Art Rhyno <art632000@yahoo.ca>
Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>