ACQ return-to-[search/invoice] buttons repair
With changes in place to support newer version of xulrunner, several
buttons no longer work in the staff client.  These include:

Return to Search in PO and Selection List pages
Return to Invoice in PO and Selection List pages
Return in lineitem worksheet page

This fix avoids using document.referrer (which is no longer available)
with an explicit 'source' CGI parameter, which contains the path and
params of the source page.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>