Fix epub complaint about image with extra colon
The "+" operators are typically only used to embed an image within an
ordered or unordered list to avoid breaking the indentation level or
numbering; here, we were using the "+" operators between regular
paragraphs, which doesn't make much sense. What made less sense was the
processing result, which bizarrely added an extra colon to the front of
the image path, leading to a (naturally) missing image and processing
error.

The fix is just to throw the image tag in between the paragraphs with no
special "+" operators; just blank lines.

Signed-off-by: Dan Scott <dscott@laurentian.ca>