Acq: New lineitems added to existing POs should inherit provider
Jeremy Miller of the Albany Public Library in Albany, OR, reports that
through the interface at
    Actions For This Record -> View/Place Orders
in the staff client, when you add a lineitem to an existing PO, the new
lineitem has a null value in the "provider" field, instead of inheriting
the expected value from the PO.

This causes problems for some Acquisitions workflows.

The feature containing the bug appeared in 2.2, and this fix for it can be
backported that far.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>